feat: update react sdk to depend on core lib
