+++ packages/core/package.json
@@ -4,9 +4,14 @@
   "description": "Pantheon Content Cloud SDK Core",
   "version": "0.0.1",
   "license": "MIT",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "build": "tsup",
+    "dev": "tsup --watch",
     "lint": "eslint . && prettier --check . && tsc --noEmit",
     "lint:fix": "eslint . --fix && prettier --write . && tsc --noEmit"
   },
@@ -19,6 +24,7 @@
   "devDependencies": {
     "@types/node": "^20.4.5",
     "eslint": "^8.46.0",
-    "eslint-config-pcc-custom": "workspace:^"
+    "eslint-config-pcc-custom": "workspace:*",
+    "tsup": "^7.1.0"
   }
 }
