+++ packages/react-sdk/src/hooks/use-articles.ts
@@ -1,9 +1,9 @@
-import { useQuery } from "../lib/apollo-client";
+import { useQuery } from "@apollo/client/react/hooks/useQuery.js";
+import { LIST_ARTICLES_QUERY } from "@pantheon-systems/pcc-sdk-core";
 import {
   ArticleQueryArgs,
   ArticleSearchArgs,
   convertSearchParamsToGQL,
-  LIST_ARTICLES_QUERY,
 } from "../lib/articles";
 import { ArticleWithoutContent } from "../types";
 
@@ -11,10 +11,14 @@ type ListArticlesResponse = {
   articles: ArticleWithoutContent[];
 };
 
+type Return = ReturnType<typeof useQuery<ListArticlesResponse>> & {
+  articles: ArticleWithoutContent[] | undefined;
+};
+
 export const useArticles = (
   args?: ArticleQueryArgs,
   searchParams?: ArticleSearchArgs,
-) => {
+): Return => {
   const queryData = useQuery<ListArticlesResponse>(LIST_ARTICLES_QUERY, {
     variables: Object.assign({}, args, convertSearchParamsToGQL(searchParams)),
   });
