+++ packages/react-sdk/src/index.ts
@@ -1,5 +1,5 @@
 export { PantheonProvider } from "./core/pantheon-context";
-export { PantheonClient } from "./core/pantheon-client";
+export { PantheonClient } from "@pantheon-systems/pcc-sdk-core";
 
 // Would be nice to have this as a separate entry point but
 // context is defined at module scope so can't be shared between
