+++ packages/react-sdk/src/lib/metadata/index.ts
@@ -2,8 +2,7 @@
  * Static helper functions for site-wide metadata.
  */
 
-import { PantheonClient } from "../../core/pantheon-client";
-import { gql } from "../apollo-client";
+import { gql, PantheonClient } from "@pantheon-systems/pcc-sdk-core";
 
 export const LIST_TAGS_QUERY = gql`
   query ListTags($id: String!) {
